MGB-17 | Основные наработки фронт + бэк

+++ meighen_bank_presenter/src/main/java/io/meighen_bank_presenter/entity/Card.java
@@ -29,7 +29,7 @@ public class Card extends BaseEntity {
     protected boolean isOtherCard;
     protected String cardIssuerName;
 
-    protected int balance;
+    protected double balance;
 
     @OneToOne
     protected CardStatus cardStatus;
